backend/requirements.txt
# ============================================================
# DecisionDNA — Multi-Agent Backend
# ============================================================
# Core Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.0

# LangChain / LangGraph (agent orchestration)
langchain==0.3.14
langchain-openai==0.3.0
langchain-community==0.3.14
langgraph==0.2.60
langchain-core==0.3.28

# OpenAI
openai==1.59.7
tiktoken==0.8.0

# Vector Store — ChromaDB
chromadb==0.6.3

# Graph Database — Neo4j
neo4j==5.27.0

# GraphRAG utilities
networkx==3.4.2
scipy==1.15.0
numpy==2.2.1
scikit-learn==1.6.1

# Embeddings
sentence-transformers==3.3.1

# Task Queue
celery==5.4.0
redis==5.2.1
kombu==5.4.2

# Data Ingestion
slack-sdk==3.34.0
google-api-python-client==2.157.0
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0

# Document Parsing
pdfplumber==0.11.4
python-docx==1.1.2
pypdf==5.1.0
unstructured==0.16.12
beautifulsoup4==4.12.3

# HTTP / Async
httpx==0.28.1
aiohttp==3.11.11
tenacity==9.0.0

# Utilities
python-multipart==0.0.20
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PyYAML==6.0.2
rich==13.9.4
loguru==0.7.3
